requirements.txt
flask
flask-session
openai
python-dotenv
markdown